refactor: storybook states higher up the tree
Mock states defined at the screen level for LoginScreen.
