+++ imdone-tasks.md
@@ -1,9 +1,4 @@
-[# Add `LoginForm` tests for the other story states](#DOING:-15)
-<!--
-created:2021-12-05T07:10:44.680Z
-TODO:2021-12-05T07:11:14.476Z
-DOING:2021-12-05T08:21:12.890Z
--->
+
 
 [# Fix Storybook stories](#DONE:-30)
 States of the stories are now wonky after React Query and MSW implementations
